backend/requirements.txt
Django>=5.0,<5.1
djangorestframework>=3.16,<3.17
psycopg2-binary>=2.9.9,<3.0
python-dotenv>=1.0.0
dj-database-url>=2.1.0
pytest>=7.4.0
pytest-django>=4.5.2
black>=23.7.0
isort>=5.12.0
flake8>=6.1.0
pre-commit>=3.3.3
factory-boy>=3.3.0
model-bakery>=1.12.0
httpx>=0.24.1
pytest-cov>=4.1.0
django-cors-headers>=4.2.0
django-filter>=23.2
sentry-sdk>=1.29.2
